Made repository pattern and DI in constructors.
I deleted DBFacade and moved methods into CheepService as DBFacade was acting as a Service for cheeps. I removed AuthorRepository as it was not used, and Author is contained within Cheep.

ConnectionString was added to AppSettings.json.

Created CheepDTO. I reused the CheepDTO in the View and removed the CheepViewModel record, as they contained same information.

Changed methods in CheepRepository to select and convert data from Cheep with predefined types into primitive types for DTO.

Removed logic that creates a database based on environment variable, as appsettings permanently creates database in working directory.